Abstract away the rendering to use simpler immediate-mode interface
Ported from my Java implementation in `gridmap-slam-robot` repo. Perhaps
a bit too verbose as of now, but more work will be done to clean up the
interface and the internals. Currently the biggest problem is that the
drawing / inserting of vertices requires access to the GL context in
case a draw call has to be made. Instead, we could buffer the vertices
and store which draw calls that need to be made and for what part of the
buffer. Depending on how many vertices we need to store, this might
use a lot of memory... Let's see how it works out!