Play around with the `Node` concept, parse Neato Lidar frames and use serial port
Just to try out the interface and how difficult/easy it is to use and
interact with other libraries and devices, create UI:s etc. Also added
some binary dumps of scan data that can be loaded and visualized.

Note that the neato lidar packet parsing is quite rough and will need
some polishing ;)