Make pubsub threaded
Since the pubsub was only ticked on UI redraws, messages were not
propagated when for example the mouse wasn't moving. This new approach
launches a background thread and uses a channel to signal that data
has been posted such that the thread can sleep until there is actual
work to do.