Make PubSub notify the app and repaint the UI when new messages are available
Temporary fix for the fact that no messages are processed unless the
UI is redrawn, i.e., the window has input events etc. to process.

Now, when notified by the PubSub framework, a UI repaint is requested.
It is not a full solution to the bigger problem (that data is processed
on the UI thread), but a step in the right direction.