Add configuration file and convert simulator to use config
This removes the function `new` in the `Node` trait as it is not really
necessary anymore, and most likely never was actually. Thus a lot of
`Node` implementations had their `new` function moved to a separate
`impl` block.

The configuration file allows to setup and configure the active nodes
in the system. Makes it easier to test different setups and parameter
variations etc. Hopefully atleast... :)