Add splitter node and make simulator more flexible
Now, the simulator always publishes tuples of observations and odometry
and there is a separate node that only splits them into separate topics
(for use with visualizing and ICP, for example).

The simulator also has the topics as optional, such that one can choose
what type of sensor to enable: scanner and/or landmark sensor.